End of session 3.2

+++ session-2/task2/ptr_size.c
@@ -8,7 +8,7 @@
 
 int main( void ) {
 
-    int *testIntPtr;
+    double *testIntPtr;
 
     printf("Int pointer size is %ld bytes\n",sizeof(testIntPtr));
